[core-xml] Adding RN support via fast-xml-parser
