--- sdk/core/core-xml/src/xml-react-native.mts
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-export * from "./xml.node.js";
